# perf: Speed up density calculus by using grid index