--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -92,6 +92,7 @@
    * CHANGED: pyvalhalla-git PyPI repository to pyvalhalla-weekly [#5310](https://github.com/valhalla/valhalla/pull/5310)
    * ADDED: `valhalla_service` to Linux Python package [#5315](https://github.com/valhalla/valhalla/pull/5315)
    * CHANGED: add full version string with git hash to any program's `--help` message [#5317](https://github.com/valhalla/valhalla/pull/5317)
+   * CHANGED: Speed up density calculus by using grid index [#5328](https://github.com/valhalla/valhalla/pull/5328)
 
 ## Release Date: 2024-10-10 Valhalla 3.5.1
 * **Removed**