# Issue

As noted in valhalla/valhalla#5201, the "enhance" stage of the graph-building process takes up about 25% of the total planet build time. Of that, the `GetDensity()` function alone accounts for around 80%. Improving `GetDensity()` would therefore significantly reduce the overall build time.

The density computation involves a large number of redundant calculations. For every node in a tile, the algorithm iterates over all neighboring nodes within a 2 km radius. To find these neighbors, it actually iterates over all nodes in the tile and often in adjacent tiles as well. For each node found within this radius, the lengths of all relevant edges are summed to produce a density value:

```cpp
  for (const auto t : tilelist) {
    // some node reading mystery from tile `t`
    for (auto node = start_node; node < end_node; ++node) {
      // Check if within radius
      if (approximator.DistanceSquared(node->latlng(base_ll)) < mr2) {
        // Get all directed edges and add length
        const DirectedEdge* directededge = newtile->directededge(node->edge_index());
        for (uint32_t i = 0; i < node->edge_count(); i++, directededge++) {
          // Exclude non-roads (parking, walkways, ferries, construction, etc.)
          if (directededge->is_road() || directededge->use() == Use::kRamp ||
              directededge->use() == Use::kTurnChannel || directededge->use() == Use::kAlley ||
              directededge->use() == Use::kEmergencyAccess) {
            roadlengths += directededge->length();
          }
        }
      }
    }
  }
```

This process repeats the same work across nodes. Since "enhance" runs in parallel per tile, we can safely use an in-memory, per-tile cache to avoid recomputing.

However, the most expensive part isn't summing edges, but the call to `approximator.DistanceSquared()`, which accounts for roughly 45% of the total time spent in `GetDensity()` (as measured on my M3 Max). So we need to optimize both the edge length lookup and the neighbor search.

Naturally, this leads to the idea of grouping values using a spatial grid. For each cell in this grid, we could precompute the `roadlengths` of all nodes within it. Then the algorithm can be redesigned as follows: for each node, find all grid cells covered by the radius and sum the values from those cells. The grid size should be large enough to reduce redundant computations but small enough to produce meaningful results with a 2 km radius.

To do this efficiently, we need fast mapping from lat/lon to grid cells and quick access to neighboring cells. A hash map like `unordered_map<CellId, float>` is the common way to represent such grid.

There are two possible approaches:

- If `cell_id` is local to each tile, then nodes on tile edges must be handled slightly differently. Additionally, up to four neighboring tiles might need to be checked. This approach benefits from the format where each node stores its coordinate using 22 + 4 bits per component.

- If `cell_id` is global (not local to a tile), then a single lookup is sufficient. The only nuance here is the need to gather adjacent nodes (up to 2 km away) from neighboring tiles. This approach has a slight performance hit due to the overhead of converting each node's coordinates into global Mercator space.

This PR explores the second approach, as it leads to a simpler (though potentially slightly slower) implementation.

```cpp
// Struct to be used as key type in hash map
struct CellId {
  uint32_t id_;

  CellId(const PointLL& ll) {
    constexpr uint32_t kGridLevel = 18;

    uint32_t x = static_cast<uint32_t>((ll.lng() + 180.0) * 1e7);
    uint32_t y = static_cast<uint32_t>((ll.lat() + 90.0) * 1e7);

    x = x >> (32 - kGridLevel);
    y = y >> (32 - kGridLevel);

    id_ = (x & 0x00FF) | ((y & 0x00FF) << 16);
  }

  CellId(uint32_t id) : id_(id) {} // for use in `neighbors()`

  // Neighboring cells ids within the given radius in degrees
  std::vector<CellId> neighbors(float radius_lng_deg, float radius_lat_deg) const {
    const int x_neighbors = std::ceil(radius_lng_deg / DensityCellId::kSizeDeg);
    const int y_neighbors = std::ceil(radius_lat_deg / DensityCellId::kSizeDeg);

    // Convert cell_id back to x, y coordinates
    const int x = id_ & 0x00FF;
    const int y = (id_ >> 16) & 0x00FF;

    std::vector<CellId> neighbors;
    neighbors.reserve(4 * x_neighbors * y_neighbors);
    for (int nx = -x_neighbors; nx <= x_neighbors; nx++) {
      for (int ny = -y_neighbors; ny <= y_neighbors; ny++) {
        // Keep only neighbors within the radius
        if (nx * nx + ny * ny <= x_neighbors * y_neighbors) {
          uint32_t neighbor_id = ((x + nx) & 0x00FF) | (((y + ny) & 0x00FF) << 16);
          neighbors.push_back(neighbor_id);
        }
      }
    }
    return neighbors;
  }
};
```

The last optimisation piece of this PR lies in the fact that nearby nodes will likely have similar density scores. Instead of computing density for every node individually, we can build a second grid that stores density per cell. Then, for each node, we simply assign the density from its corresponding cell.

## Timings

*measured on M3 Max using europe-latest.osm.pbf*

master:

```
$ ./valhalla_build_tiles -c ./valhalla_data/valhalla.json ~/europe-latest.osm.pbf | rg TIMING
2025/06/21 11:57:39.443813 [TIMING] src/mjolnir/pbfgraphparser.cc::clarify_and_fix took 162s
2025/06/21 11:57:44.295937 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseWays took 526s
2025/06/21 11:58:21.534922 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseRelations took 37s
2025/06/21 12:14:48.855632 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseNodes took 987s
2025/06/21 12:19:22.058906 [TIMING] src/mjolnir/graphbuilder.cc::BuildEdges took 273s
2025/06/21 12:21:09.999029 [TIMING] src/mjolnir/linkclassification.cc::ReclassifyLinks took 107s
2025/06/21 12:22:31.685841 [TIMING] src/mjolnir/ferry_connections.cc::ReclassifyFerryConnections took 81s
2025/06/21 12:32:00.956373 [TIMING] src/mjolnir/graphbuilder.cc::BuildLocalTiles took 569s
2025/06/21 12:32:00.956392 [TIMING] src/mjolnir/graphbuilder.cc::Build took 758s
2025/06/21 13:13:50.237525 [TIMING] src/mjolnir/graphenhancer.cc::Enhance took 2509s
2025/06/21 13:13:50.238947 [TIMING] src/mjolnir/graphfilter.cc::Filter took 0s
2025/06/21 13:14:23.877366 [TIMING] src/mjolnir/hierarchybuilder.cc::CreateNodeAssociations took 33s
2025/06/21 13:14:49.692481 [TIMING] src/mjolnir/hierarchybuilder.cc::SortSequences took 25s
2025/06/21 13:23:09.851717 [TIMING] src/mjolnir/hierarchybuilder.cc::FormTilesInNewLevel took 500s
2025/06/21 13:23:11.407519 [TIMING] src/mjolnir/hierarchybuilder.cc::RemoveUnusedLocalTiles took 1s
2025/06/21 13:23:11.432481 [TIMING] src/mjolnir/hierarchybuilder.cc::Build took 561s
2025/06/21 13:24:23.336954 [TIMING] src/mjolnir/shortcutbuilder.cc::Build took 71s
2025/06/21 13:27:48.341543 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 21s
2025/06/21 13:29:23.041126 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 62s
2025/06/21 13:30:17.807519 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 41s
2025/06/21 13:30:17.808810 [TIMING] src/mjolnir/restrictionbuilder.cc::Build took 354s
2025/06/21 13:33:38.921170 [TIMING] src/mjolnir/graphvalidator.cc::Validate took 201s
2025/06/21 13:33:39.347315 [TIMING] src/mjolnir/osmdata.cc::cleanup_temp_files took 0s
2025/06/21 13:33:40.419637 [TIMING] src/mjolnir/util.cc::build_tile_set took 6282s
```

this PR

```
$ ./valhalla_build_tiles -c ./valhalla_data/valhalla.json ~/europe-latest.osm.pbf | rg TIMING
2025/06/22 00:18:26.058923 [TIMING] src/mjolnir/pbfgraphparser.cc::clarify_and_fix took 181s
2025/06/22 00:18:30.984076 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseWays took 551s
2025/06/22 00:19:07.546185 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseRelations took 36s
2025/06/22 00:35:42.596176 [TIMING] src/mjolnir/pbfgraphparser.cc::ParseNodes took 994s
2025/06/22 00:40:27.279595 [TIMING] src/mjolnir/graphbuilder.cc::BuildEdges took 284s
2025/06/22 00:42:15.678095 [TIMING] src/mjolnir/linkclassification.cc::ReclassifyLinks took 108s
2025/06/22 00:43:36.586335 [TIMING] src/mjolnir/ferry_connections.cc::ReclassifyFerryConnections took 80s
2025/06/22 00:52:59.256056 [TIMING] src/mjolnir/graphbuilder.cc::BuildLocalTiles took 562s
2025/06/22 00:52:59.256072 [TIMING] src/mjolnir/graphbuilder.cc::Build took 751s
2025/06/22 01:02:13.835846 [TIMING] src/mjolnir/graphenhancer.cc::Enhance took 554s
2025/06/22 01:02:13.835874 [TIMING] src/mjolnir/graphfilter.cc::Filter took 0s
2025/06/22 01:02:48.303948 [TIMING] src/mjolnir/hierarchybuilder.cc::CreateNodeAssociations took 34s
2025/06/22 01:03:13.357186 [TIMING] src/mjolnir/hierarchybuilder.cc::SortSequences took 25s
2025/06/22 01:11:13.052308 [TIMING] src/mjolnir/hierarchybuilder.cc::FormTilesInNewLevel took 479s
2025/06/22 01:11:14.591211 [TIMING] src/mjolnir/hierarchybuilder.cc::RemoveUnusedLocalTiles took 1s
2025/06/22 01:11:14.605179 [TIMING] src/mjolnir/hierarchybuilder.cc::Build took 540s
2025/06/22 01:12:24.392332 [TIMING] src/mjolnir/shortcutbuilder.cc::Build took 69s
2025/06/22 01:15:44.979207 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 21s
2025/06/22 01:17:17.897876 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 61s
2025/06/22 01:18:11.031658 [TIMING] src/mjolnir/restrictionbuilder.cc::HandleOnlyRestrictionProperties took 40s
2025/06/22 01:18:11.032015 [TIMING] src/mjolnir/restrictionbuilder.cc::Build took 346s
2025/06/22 01:21:31.356295 [TIMING] src/mjolnir/graphvalidator.cc::Validate took 200s
2025/06/22 01:21:31.807840 [TIMING] src/mjolnir/osmdata.cc::cleanup_temp_files took 0s
2025/06/22 01:21:32.887331 [TIMING] src/mjolnir/util.cc::build_tile_set took 4333s
```

With noteceable improvement for `src/mjolnir/graphenhancer.cc::Enhance` from `2509s` to `554s`

## Obserable density changes

| Location | master | this PR |
| --- | --- | --- |
| Berlin | <img width="1188" alt="image" src="https://github.com/user-attachments/assets/17cf681e-f6ba-4a9c-9bea-86873af99b6e" /> | <img width="1189" alt="image" src="https://github.com/user-attachments/assets/99c1bec5-352a-46bd-9bdc-460c291360df" /> |
| Paris | <img width="1190" alt="image" src="https://github.com/user-attachments/assets/edf0510e-f162-4c51-ac19-78e440db7e54" /> |  <img width="1186" alt="image" src="https://github.com/user-attachments/assets/de031e0e-b1c0-427b-a52e-2e459e8674e3" /> |
| Köpenhamn | <img width="1190" alt="image" src="https://github.com/user-attachments/assets/4b156804-8a19-4238-8f58-672ce1943789" /> | <img width="1191" alt="image" src="https://github.com/user-attachments/assets/ac50c87f-6992-4114-97cc-9fdd3cac21e6" /> |
| Stockholm | <img width="1190" alt="image" src="https://github.com/user-attachments/assets/ea1c7722-d0ab-4df1-9b41-998e40cf264d" /> | <img width="1190" alt="image" src="https://github.com/user-attachments/assets/22ae98ae-13d8-489b-bbda-763ed1350348" /> |

*density values are mapped from `[0..15]` to black-red-yellow-green palette*

## Tasklist

 - [ ] Add tests
 - [x] Add #fixes with the issue number that this PR addresses
 - [ ] Update the docs with any new request parameters or changes to behavior described
 - [x] Update the [changelog](CHANGELOG.md)
 - [ ] If you made changes to the lua files, update the [taginfo](taginfo.json) too.

## Requirements / Relations

This idea was mentioned in the first time here - https://github.com/valhalla/valhalla/pull/5186#issuecomment-2847944742

closes https://github.com/valhalla/valhalla/issues/5201