@kinkard 2025-06-22
```
[FAIL] isochrone
make[3]: *** [test/isochrone.log] Error 1
[==========] Running 9 tests from 1 test suite.
make[3]: *** Deleting file `test/isochrone.log'
[----------] Global test environment set-up.
make[2]: *** [test/CMakeFiles/run-isochrone.dir/all] Error 2
[----------] 9 tests from Isochrones
make[2]: *** Waiting for unfinished jobs....
[ RUN      ] Isochrones.Basic
/Users/runner/work/valhalla/valhalla/test/isochrone.cc:60: Failure
Value of: test::shape_equality(actual, expected, 41)
  Actual: false (shape exceeds tolerance by 117.6858309[660](https://github.com/valhalla/valhalla/actions/runs/15806577209/job/44552191999?pr=5328#step:13:661)3851)
Expected: true
Google Test trace:
/Users/runner/work/valhalla/valhalla/test/isochrone.cc:189: basic request 1 failed

/Users/runner/work/valhalla/valhalla/test/isochrone.cc:60: Failure
Value of: test::shape_equality(actual, expected, 41)
  Actual: false (shape exceeds tolerance by 18.778767575439332)
Expected: true
Google Test trace:
/Users/runner/work/valhalla/valhalla/test/isochrone.cc:227: basic request 5 failed

[  FAILED  ] Isochrones.Basic (146 ms)
```

I would highly appreciate any help with this failing test case